+++ packages/expo-cli/src/CommandError.ts
@@ -1,5 +1,3 @@
-import ExtendableError from 'es6-error';
-
 const ERROR_PREFIX = 'Error: ';
 
 export const ErrorCodes = {
@@ -21,24 +19,23 @@ export type ErrorCode = keyof typeof ErrorCodes;
 /**
  * General error, formatted as a message in red text when caught by expo-cli (no stack trace is printed). Should be used in favor of `log.error()` in most cases.
  */
-export default class CommandError extends ExtendableError {
+export default class CommandError extends Error {
+  name = 'CommandError';
+  readonly isCommandError = true;
   code: string;
-  isCommandError: true;
 
   constructor(code: string, message: string = '') {
+    super('');
     // If e.toString() was called to get `message` we don't want it to look
     // like "Error: Error:".
     if (message.startsWith(ERROR_PREFIX)) {
       message = message.substring(ERROR_PREFIX.length);
     }
 
-    super(message || code);
-
+    this.message = message || code;
     this.code = code;
-    this.isCommandError = true;
   }
 }
-
 export class AbortCommandError extends CommandError {
   constructor() {
     super('ABORTED', 'Interactive prompt was cancelled.');