+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This is the log of notable changes to Expo CLI and related packages.
 
 ### 🛠 Breaking changes
 
+- [cli] Removed `generate-module` command ([#2903](https://github.com/expo/expo-cli/pull/2903))
+
 ### 🎉 New features
 
 ### 🐛 Bug fixes